+++ core/base/Module.php
@@ -2,11 +2,11 @@
 
 namespace luya\base;
 
-use luya\theme\Theme;
 use Yii;
 use yii\base\InvalidConfigException;
 use luya\console\interfaces\ImportControllerInterface;
 use luya\helpers\ObjectHelper;
+use yii\base\Application;
 
 /**
  * LUYA Module base class.
@@ -198,6 +198,22 @@ abstract class Module extends \yii\base\Module
     }
 
     /**
+     * The LUYA Bootstrap method will be invoken when the application starts.
+     *
+     * As LUYA modules will be loaded while bootstraping, this method will ALWAYS be invoken when
+     * the application starts.
+     *
+     * Compared to the {{yii\base\BootstrapInterface}} the module or class must still be configured
+     * to bootstrap in the configuration section, the {{luyaBootstrap}} will be invoken always.
+     *
+     * @param Application $app
+     * @since 1.0.21
+     */
+    public function luyaBootstrap(Application $app)
+    {
+    }
+
+    /**
      * Override the default implementation of Yii's getLayoutPath(). If the property `$useAppLayoutPath` is true,
      * the *@app* namespace views will be looked up for view files.
      * Else the layout path of the active theme will be used.
@@ -207,7 +223,7 @@ abstract class Module extends \yii\base\Module
      */
     public function getLayoutPath()
     {
-        if (Yii::$app->themeManager->hasActiveTheme && $this->useAppLayoutPath){
+        if (Yii::$app->themeManager->hasActiveTheme && $this->useAppLayoutPath) {
             $this->setLayoutPath(Yii::$app->themeManager->activeTheme->layoutPath);
         } elseif ($this->useAppLayoutPath) {
             $this->setLayoutPath('@app/views/'.$this->id.'/layouts');